extract op
